Move posts-related selectors under state/posts
