+++ client/my-sites/post-type-list/post-actions-ellipsis-menu/duplicate.jsx
@@ -12,7 +12,7 @@ import { includes } from 'lodash';
  */
 import PopoverMenuItem from 'components/popover/menu-item';
 import { getPost } from 'state/posts/selectors';
-import canCurrentUserEditPost from 'state/selectors/can-current-user-edit-post';
+import { canCurrentUserEditPost } from 'state/posts/selectors/can-current-user-edit-post';
 import { getEditorDuplicatePostPath } from 'state/editor/selectors';
 import { isJetpackSite } from 'state/sites/selectors';
 import QueryJetpackModules from 'components/data/query-jetpack-modules';
