+++ client/post-editor/editor-revisions-list/view-buttons.jsx
@@ -14,7 +14,7 @@ import {
 	splitPostRevisionsDiffView,
 	unifyPostRevisionsDiffView,
 } from 'state/posts/revisions/actions';
-import getPostRevisionsDiffView from 'state/selectors/get-post-revisions-diff-view';
+import { getPostRevisionsDiffView } from 'state/posts/selectors/get-post-revisions-diff-view';
 
 const EditorRevisionsListViewButtons = ( { diffView, viewSplit, viewUnified } ) => {
 	const translate = useTranslate();
