+++ client/state/posts/selectors/count-post-likes.js
@@ -9,13 +9,13 @@ import { get } from 'lodash';
 import 'state/posts/init';
 
 /**
- * Returns the post likes for a given site ID, post ID.
+ * Returns the total of post likes for a given site ID, post ID.
  *
  * @param  {object}  state  Global state tree
  * @param  {number}  siteId Site ID
  * @param  {number}  postId Post ID
  * @returns {Array}          Post Likes
  */
-export default function getPostLikes( state, siteId, postId ) {
-	return get( state.posts.likes.items, [ siteId, postId, 'likes' ], null );
+export function countPostLikes( state, siteId, postId ) {
+	return get( state.posts.likes.items, [ siteId, postId, 'found' ], null );
 }
