+++ client/state/posts/selectors/get-post-like-last-updated.js
@@ -8,6 +8,6 @@ import { get } from 'lodash';
  */
 import 'state/posts/init';
 
-export default function getPostLikeLastUpdated( state, siteId, postId ) {
+export function getPostLikeLastUpdated( state, siteId, postId ) {
 	return get( state.posts.likes.items, [ siteId, postId, 'lastUpdated' ] );
 }
