+++ client/state/posts/selectors/get-post-revision.js
@@ -10,10 +10,8 @@ import createSelector from 'lib/create-selector';
 
 import 'state/posts/init';
 
-const getPostRevision = createSelector(
+export const getPostRevision = createSelector(
 	( state, siteId, postId, revisionId ) =>
 		get( state.posts.revisions.diffs, [ siteId, postId, 'revisions', revisionId ], null ),
 	( state ) => [ state.posts.revisions.diffs ]
 );
-
-export default getPostRevision;
