+++ client/state/posts/selectors/get-post-revisions-diff.js
@@ -8,13 +8,7 @@ import { get } from 'lodash';
  */
 import 'state/posts/init';
 
-export default function getPostRevisionsDiff(
-	state,
-	siteId,
-	postId,
-	fromRevisionId,
-	toRevisionId
-) {
+export function getPostRevisionsDiff( state, siteId, postId, fromRevisionId, toRevisionId ) {
 	const key = `${ fromRevisionId || 0 }:${ toRevisionId || 0 }`;
 	return get( state.posts.revisions.diffs, [ siteId, postId, key, 'diff' ], {} );
 }
