+++ client/state/posts/selectors/get-post-revisions-selected-revision.js
@@ -3,10 +3,10 @@
  */
 import { getSelectedSiteId } from 'state/ui/selectors';
 import { getEditorPostId } from 'state/editor/selectors';
-import getPostRevision from 'state/selectors/get-post-revision';
-import getPostRevisionsSelectedRevisionId from 'state/selectors/get-post-revisions-selected-revision-id';
+import { getPostRevision } from 'state/posts/selectors/get-post-revision';
+import { getPostRevisionsSelectedRevisionId } from 'state/posts/selectors/get-post-revisions-selected-revision-id';
 
-export default function getPostRevisionsSelectedRevision( state ) {
+export function getPostRevisionsSelectedRevision( state ) {
 	const siteId = getSelectedSiteId( state );
 	const postId = getEditorPostId( state );
 	const revisionId = getPostRevisionsSelectedRevisionId( state );
