+++ client/state/posts/selectors/index.js
@@ -1,10 +1,24 @@
+export { canCurrentUserEditPost } from 'state/posts/selectors/can-current-user-edit-post';
+export { countPostLikes } from 'state/posts/selectors/count-post-likes';
+export { editedPostHasContent } from 'state/posts/selectors/edited-post-has-content';
 export { getEditedPost } from 'state/posts/selectors/get-edited-post';
 export { getEditedPostSlug } from 'state/posts/selectors/get-edited-post-slug';
 export { getEditedPostValue } from 'state/posts/selectors/get-edited-post-value';
 export { getNormalizedPost } from 'state/posts/selectors/get-normalized-post';
 export { getPost } from 'state/posts/selectors/get-post';
 export { getPostEdits } from 'state/posts/selectors/get-post-edits';
+export { getPostLikeCount } from 'state/posts/selectors/get-post-like-count';
+export { getPostLikeLastUpdated } from 'state/posts/selectors/get-post-like-last-updated';
+export { getPostLikes } from 'state/posts/selectors/get-post-likes';
 export { getPostPreviewUrl } from 'state/posts/selectors/get-post-preview-url';
+export { getPostRevision } from 'state/posts/selectors/get-post-revision';
+export { getPostRevisions } from 'state/posts/selectors/get-post-revisions';
+export { getPostRevisionsAuthorsId } from 'state/posts/selectors/get-post-revisions-authors-id';
+export { getPostRevisionsComparisons } from 'state/posts/selectors/get-post-revisions-comparisons';
+export { getPostRevisionsDiff } from 'state/posts/selectors/get-post-revisions-diff';
+export { getPostRevisionsDiffView } from 'state/posts/selectors/get-post-revisions-diff-view';
+export { getPostRevisionsSelectedRevision } from 'state/posts/selectors/get-post-revisions-selected-revision';
+export { getPostRevisionsSelectedRevisionId } from 'state/posts/selectors/get-post-revisions-selected-revision-id';
 export { getPostsForQuery } from 'state/posts/selectors/get-posts-for-query';
 export { getPostsForQueryIgnoringPage } from 'state/posts/selectors/get-posts-for-query-ignoring-page';
 export { getPostsFoundForQuery } from 'state/posts/selectors/get-posts-found-for-query';
@@ -15,7 +29,9 @@ export { getSitePostsByTerm } from 'state/posts/selectors/get-site-posts-by-term
 export { isEditedPostDirty } from 'state/posts/selectors/is-edited-post-dirty';
 export { isEditedPostPasswordProtected } from 'state/posts/selectors/is-edited-post-password-protected';
 export { isEditedPostPasswordProtectedWithValidPassword } from 'state/posts/selectors/is-edited-post-password-protected-with-valid-password';
+export { isLikedPost } from 'state/posts/selectors/is-liked-post';
 export { isPostPublished } from 'state/posts/selectors/is-post-published';
+export { isPostRevisionsDialogVisible } from 'state/posts/selectors/is-post-revisions-dialog-visible';
 export { isPostsLastPageForQuery } from 'state/posts/selectors/is-posts-last-page-for-query';
 export { isRequestingPostsForQuery } from 'state/posts/selectors/is-requesting-posts-for-query';
 export { isRequestingPostsForQueryIgnoringPage } from 'state/posts/selectors/is-requesting-posts-for-query-ignoring-page';
