+++ client/state/selectors/test/count-post-likes.js
@@ -6,7 +6,7 @@ import { expect } from 'chai';
 /**
  * Internal dependencies
  */
-import countPostLikes from 'state/selectors/count-post-likes';
+import { countPostLikes } from 'state/posts/selectors/count-post-likes';
 
 describe( 'countPostLikes()', () => {
 	test( 'should return null if the site has never been fetched', () => {
