+++ client/state/selectors/test/get-post-revision.js
@@ -6,7 +6,7 @@ import { expect } from 'chai';
 /**
  * Internal dependencies
  */
-import getPostRevision from 'state/selectors/get-post-revision';
+import { getPostRevision } from 'state/posts/selectors/get-post-revision';
 
 describe( 'getPostRevision', () => {
 	test( 'should return `null` if there is no revision in the state for `siteId, postId`', () => {
